refactor: Fixed the fileFullPath error and used a new 'toolkit' folder.
The bug was that previous versions still hung around and therefore all of those old .CAB files were bieng pased to the second Get-ChocolateyUnzip. I refactored the code in here to only use the newest .CAB file found and also to keep all of the extracted files in a toolkit folder to allow easier identification and removal. Finally I removed the initial extracted folder that contained the .CAB file.